+++ StripeFinancialConnections/StripeFinancialConnections/Source/Native/NetworkingLinkLoginWarmup/NetworkingLinkLoginWarmupViewController.swift
@@ -20,14 +20,17 @@ protocol NetworkingLinkLoginWarmupViewControllerDelegate: AnyObject {
     func networkingLinkLoginWarmupViewController(_ viewController: NetworkingLinkLoginWarmupViewController, didReceiveTerminalError error: Error)
 }
 
-final class NetworkingLinkLoginWarmupViewController: UIViewController {
+final class NetworkingLinkLoginWarmupViewController: SheetViewController {
 
     private let dataSource: NetworkingLinkLoginWarmupDataSource
     weak var delegate: NetworkingLinkLoginWarmupViewControllerDelegate?
 
-    init(dataSource: NetworkingLinkLoginWarmupDataSource) {
+    init(
+        dataSource: NetworkingLinkLoginWarmupDataSource,
+        panePresentationStyle: PanePresentationStyle
+    ) {
         self.dataSource = dataSource
-        super.init(nibName: nil, bundle: nil)
+        super.init(panePresentationStyle: panePresentationStyle)
     }
 
     required init?(coder: NSCoder) {
@@ -36,32 +39,49 @@ final class NetworkingLinkLoginWarmupViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.backgroundColor = .customBackgroundColor
-
-        let pane = PaneWithHeaderLayoutView(
-            title: STPLocalizedString(
-                "Sign in to Link",
-                "The title of a screen where users are informed that they can sign-in-to Link."
+        setup(
+            withContentView: PaneLayoutView.createContentView(
+                iconView: RoundedIconView(
+                    image: .image(.person),
+                    style: .circle
+                ),
+                title: STPLocalizedString(
+                    "Continue with Link",
+                    "The title of a screen where users are informed that they can sign-in-to Link."
+                ),
+                subtitle: STPLocalizedString(
+                    "Use information you previously saved with your Link account.",
+                    "The subtitle/description of a screen where users are informed that they can sign-in-to Link."
+                ),
+                contentView: NetworkingLinkLoginWarmupBodyView(
+                    // `accountholderCustomerEmailAddress` should always be non-null, and
+                    // since the email is only used as a visual, it's not worth to throw an error
+                    // if it is null
+                    email: dataSource.manifest.accountholderCustomerEmailAddress ?? "you"
+                )
             ),
-            subtitle: STPLocalizedString(
-                "It looks like you have a Link account. Signing in will let you quickly access your saved bank accounts.",
-                "The subtitle/description of a screen where users are informed that they can sign-in-to Link."
-            ),
-            contentView: NetworkingLinkLoginWarmupBodyView(
-                // `accountholderCustomerEmailAddress` should always be non-null, and
-                // since the email is only used as a visual, it's not worth to throw an error
-                // if it is null
-                email: dataSource.manifest.accountholderCustomerEmailAddress ?? "you",
-                didSelectContinue: { [weak self] in
-                    self?.didSelectContinue()
-                },
-                didSelectSkip: { [weak self] in
-                    self?.didSelectSkip()
-                }
-            ),
-            footerView: nil
+            footerView: PaneLayoutView.createFooterView(
+                primaryButtonConfiguration: PaneLayoutView.ButtonConfiguration(
+                    title: STPLocalizedString(
+                        "Continue with Link",
+                        "A button title. This button, when pressed, will automatically log-in the user with their e-mail to Link (one-click checkout provider)."
+                    ),
+                    accessibilityIdentifier: "link_continue_button",
+                    action: { [weak self] in
+                        self?.didSelectContinue()
+                    }
+                ),
+                secondaryButtonConfiguration: PaneLayoutView.ButtonConfiguration(
+                    title: STPLocalizedString(
+                        "Not now",
+                        "A button title. This button, when pressed, will skip logging in the user with their e-mail to Link (one-click checkout provider)."
+                    ),
+                    action: { [weak self] in
+                        self?.didSelectSkip()
+                    }
+                )
+            ).footerView
         )
-        pane.addTo(view: view)
     }
 
     private func didSelectContinue() {